Make sure LaunchAgent file is unloaded and reloaded on reset

+++ lib/pushes/launch_agent.rb
@@ -17,6 +17,7 @@ class Pushes::LaunchAgent
   end
 
   def stop
+    return unless File.exist?(PLIST_PATH)
     `launchctl unload #{PLIST_PATH}`
     FileUtils.rm(PLIST_PATH)
   end
